requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pgvector
python-multipart
Pillow
scikit-learn
